Move components into corresponding files
Move components into corresponding files
Fix missing env var for express
