+++ run.bat
@@ -1,6 +1,9 @@
 @echo off
 cd ./api
-docker build -t matching_service .
+call npm run build
+:: npm run start:dev
+
+call docker build -t matching_service .
 
 cd ..
 docker-compose up
